fix the calculation of sediment diffusion
dt/dx terms are added in the calculation of sediment diffusion in turb2d.py.
Also, when run_multiflow_script.py is run, the file name and directory name are determined in config.yml.
In addition, the image of the sediment distribution up to the specified run no. is output at the end of the calculation.
An error message is output when the sediment entrainment function is specified or misspelled, which is not implemented.