Fixed bug in boundary conidtions  and
Fixed a bug in update_boundary_conditions that caused the Kh value of the grid upstream from the downstream end to be 0.
Cf_node was defined with the value of link, but now it uses the value of node.
Fixed the termination condition of continuous flow calculation in run_flows.py.
In addition, changed to separate calculation configuration files for run_multi_flows.py and run_turb2d_script.py.