Fixed bug in feature to generate repetitive flow
Previously, run_flows.py did not reset the flow condition at the end of a single flow,
so the flow was not repeated, and the flow was continuous for(the specified number of times)*(flow duration) seconds.
Therefore, the scripts were modified to reset the flow condition at the end of each calculation when a repetitive flow is generated.
Also, typo in config_runmulti.yml was fixed.