Fixed run_flows to allow alpha_4eq to select None
alpha_4eq was randomly generated even when 3eq was selected.
Therefore, when run_multiple_flows.py is executed, it can be selected not to generate alpha_4eq.
If alpha_4eq_min and alpha_4eq_max are set to None, alpha_4eq is not generated and alpha_4eq is stored as None.