Fixed bug in _calculate_turbulent_kinetic_energy
The updated cf was not used in the calculation of the production term of turbulent kinetic energy.
Furthermore, the initial Kh values were all set to zero.
Therefore, the initial Kh was set based on the inlet conditions,
and the calculation of the production term was fixed to use the updated cf values.