Add feature of generation of nested grids
create_nested_grid function was added to utils.py so that nested grids can be generated.
The create_nested_grid function first creates a parent grid with a coarse mesh, and then performs a linear interpolation of the topographic elevation of parent grid to a child grid with a fine mesh.